third_party/WebKit/LayoutTests/inspector/elements/event-listener-sidebar-jquery2.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="resources/jquery-2.1.4.min.js"></script>
<script>
function setupEventListeners()
{
    var node = $("#node")[0];
    $("#node").click(function(){ console.log("first jquery"); });
    $("#node").click(function(){ console.log("second jquery"); });
    node.addEventListener("click", function() { console.log("addEventListener"); });
    node.onclick = function() { console.log("onclick"); }
}

function test()
{
    Common.settingForTest("showEventListenersForAncestors").set(true);
    InspectorTest.selectNodeWithId("node", step1);

    function step1()
    {
        InspectorTest.showEventListenersWidget();
        InspectorTest.expandAndDumpSelectedElementEventListeners(step2);
    }

    function step2()
    {
        InspectorTest.addResult("Remove listeners..");
        var eventListenersWidget = InspectorTest.eventListenersWidget();
        var listenerTypes = eventListenersWidget._eventListenersView._treeOutline.rootElement().children();
        var promises = [];
        for (var i in listenerTypes) {
            var listenersItems = listenerTypes[i].children();
            for (var j in listenersItems)
                promises.push(listenersItems[j].eventListener().remove());
        }
        InspectorTest.expandAndDumpSelectedElementEventListeners(InspectorTest.completeTest.bind(this));
        Promise.all(promises).then(() => eventListenersWidget.doUpdate());
    }
}


function onloadHandler()
{
    setupEventListeners();
    runTest();
}
</script>
</head>

<body onload="onloadHandler()">
<p>
Tests event listeners output in the Elements sidebar panel.
</p>
<button id="node">Inspect Me</button>
</body>
</html>
