third_party/WebKit/LayoutTests/inspector/elements/event-listener-sidebar-jquery2-expected.txt

Tests event listeners output in the Elements sidebar panel.

Inspect Me

======== click ========
== frameworkUser
[expanded] button#nodeRemoveevent-listener-sidebar-jquery2.html:11
    useCapture: true
    handler: function (){ console.log("second jquery"); }
== frameworkUser
[expanded] button#nodeRemoveevent-listener-sidebar-jquery2.html:10
    useCapture: true
    handler: function (){ console.log("first jquery"); }
== normal
[expanded] button#nodeRemoveevent-listener-sidebar-jquery2.html:13
    useCapture: false
    handler: function () { constole.log("onclick"); }
== normal
[expanded] button#nodeRemoveevent-listener-sidebar-jquery2.html:12
    useCapture: false
    handler: function () { console.log("addEventListener"); }
== frameworkInternal
[expanded] button#nodeRemovejquery-2.1.4.min.js:3
    useCapture: false
    handler: function (b){return typeof n!==U&&n.event.triggered!==b.type?n.event.dispatch.apply(a,arguments):void 0}

======== load ========
== normal
[expanded] WindowRemoveevent-listener-sidebar-jquery2.html:52
    useCapture: false
    handler: function onload(event) {
  onloadHandler()
}
Remove listeners..
